requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-filters==0.2.1
djangorestframework==3.10.3
idna==2.8
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
